+++ src/org/lockss/servlet/HashCUS.java
@@ -1,5 +1,5 @@
 /*
- * $Id: HashCUS.java,v 1.40 2009-02-05 05:08:47 tlipkis Exp $
+ * $Id: HashCUS.java,v 1.41 2009-04-07 04:51:24 tlipkis Exp $
  */
 
 /*
@@ -127,7 +127,7 @@ public class HashCUS extends LockssServlet {
 
   MessageDigest digest;
   byte[] hashResult;
-  int bytesHashed;
+  long bytesHashed;
   int filesHashed;
   boolean showResult;
   protected void resetLocals() {
@@ -346,7 +346,7 @@ public class HashCUS extends LockssServlet {
     if (verifier != null) {
       addResultRow(tbl, "Verifier", byteString(verifier));
     }
-    addResultRow(tbl, "Size", Integer.toString(bytesHashed));
+    addResultRow(tbl, "Size", Long.toString(bytesHashed));
     addResultRow(tbl, "Hash", byteString(hashResult));
 
     addResultRow(tbl, "Time", getElapsedString());
@@ -385,7 +385,7 @@ public class HashCUS extends LockssServlet {
 
     addResultRow(tbl, "CUSS", cus.getSpec().toString());
     addResultRow(tbl, "Files", Integer.toString(filesHashed));
-    addResultRow(tbl, "Size", Integer.toString(bytesHashed));
+    addResultRow(tbl, "Size", Long.toString(bytesHashed));
     addResultRow(tbl, "Time", getElapsedString());
     if (blockFile != null && blockFile.exists()) {
       tbl.newRow();
